src/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
torch==2.0.0
torchvision==0.15.1
Pillow==9.4.0
numpy==1.24.0